Use Base classes for importer and exporters.
