+++ jrnl/plugins/exporter/json_exporter.py
@@ -5,15 +5,18 @@
 
 import json
 
-from jrnl.plugins.exporter.text_exporter import Exporter as TextExporter
+from jrnl.plugins.base import BaseExporter
 from jrnl.plugins.util import get_tags_count
 
+from ... import __version__
 
-class Exporter(TextExporter):
+
+class Exporter(BaseExporter):
     """This Exporter can convert entries and journals into json."""
 
     names = ["json"]
     extension = "json"
+    version = __version__
 
     @classmethod
     def entry_to_dict(cls, entry):
