+++ jrnl/plugins/importer/jrnl_importer.py
@@ -5,17 +5,16 @@
 
 import sys
 
-from jrnl import __version__
+from jrnl.plugins.base import BaseImporter
 
-class Importer:
+from ... import __version__
+
+
+class Importer(BaseImporter):
     """This plugin imports entries from other jrnl files."""
 
     names = ["jrnl"]
     version = __version__
-
-    @classmethod
-    def class_path(cls):
-        return cls.__module__
 
     @staticmethod
     def import_(journal, input=None):
@@ -34,7 +33,9 @@ class Importer:
         journal.import_(other_journal_txt)
         new_cnt = len(journal.entries)
         print(
-            "[{} imported to {} journal]".format(new_cnt - old_cnt, journal.name),
+            "[{} entries imported to '{}' journal]".format(
+                new_cnt - old_cnt, journal.name
+            ),
             file=sys.stderr,
         )
         journal.write()
